+++ src/main/java/com/gregtechceu/gtceu/syncdata/GTRecipePayload.java
@@ -37,7 +37,7 @@ public class GTRecipePayload extends ObjectTypedPayload<GTRecipe> {
         tag.putString("id", payload.id.toString());
         tag.put("recipe",
                 GTRecipeSerializer.CODEC.encodeStart(NbtOps.INSTANCE, payload).result().orElse(new CompoundTag()));
-        tag.putInt("parallels", payload.parallels);
+        tag.putInt("parallels", (int) payload.parallels);
         tag.putInt("ocLevel", payload.ocLevel);
         return tag;
     }
@@ -74,7 +74,7 @@ public class GTRecipePayload extends ObjectTypedPayload<GTRecipe> {
     public void writePayload(FriendlyByteBuf buf) {
         buf.writeResourceLocation(this.payload.id);
         GTRecipeSerializer.SERIALIZER.toNetwork(buf, this.payload);
-        buf.writeInt(this.payload.parallels);
+        buf.writeInt((int) this.payload.parallels);
         buf.writeInt(this.payload.ocLevel);
     }
